# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: GCCS API
  version: 0.1.0
  summary: Government Contractor Compliance SaaS API
  description: >
    API specification for the GCCS MVP. GCCS is a No-CUI compliance management
    SaaS for small U.S. government contractors. The API is tenant-scoped and
    supports company profiles, contract intake, clause obligations, evidence,
    calendar tasks, CMMC readiness, subcontractor flow-downs, reports, and
    source-backed compliance content.
  contact:
    name: GCCS Product Team
servers:
  - url: http://localhost:5062
    description: Local ASP.NET Core API
  - url: https://api.gccs.example.com
    description: Production API placeholder
security:
  - bearerAuth: []
...
paths:
...
  /api/me/access:
    get:
      tags:
        - Access
      summary: Get current user roles and permissions
      operationId: getCurrentUserAccess
      description: >
        Returns role-derived permissions for the authenticated tenant context so
        clients can hide restricted actions. Server-side authorization remains
        authoritative.
      responses:
        "200":
          description: Current user access metadata.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CurrentUserAccess"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/compliance/overview:
    get:
      tags:
        - Compliance Overview
      summary: Get compliance workspace overview
      operationId: getComplianceOverview
      responses:
        "200":
          description: Dashboard overview.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ComplianceOverview"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/tenants:
    post:
      tags:
        - Tenants
      summary: Create a No-CUI tenant workspace
      operationId: createTenant
      description: >
        Creates an isolated customer tenant in the MVP No-CUI posture. Tenant
        creation requires the ManageTenant permission and creates an audit event
        in the new tenant scope.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTenantRequest"
      responses:
        "201":
          description: Created tenant.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/tenants/{tenantId}:
    get:
      tags:
        - Tenants
      summary: Get tenant metadata in the current tenant scope
      operationId: getTenant
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Tenant metadata.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/tenants/{tenantId}/status:
    patch:
      tags:
        - Tenants
      summary: Update tenant status
      operationId: updateTenantStatus
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTenantStatusRequest"
      responses:
        "200":
          description: Updated tenant.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/tenant-members:
    get:
      tags:
        - Tenants
      summary: List users assigned to the current tenant
      operationId: listCurrentTenantMembers
      description: >
        Returns only memberships for the authenticated active tenant. Requires
        the ManageUsers permission.
      responses:
        "200":
          description: Current-tenant members.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TenantMember"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      tags:
        - Tenants
      summary: Assign a user to the current tenant
      operationId: assignTenantMember
      description: >
        Adds an explicit membership between an existing or newly recorded user
        and the authenticated tenant. Duplicate user membership in the same
        tenant is rejected and successful changes are audit logged.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssignTenantMemberRequest"
      responses:
        "201":
          description: Created current-tenant membership.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenantMember"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          description: User already has a membership in the current tenant.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
  /api/tenant-members/{membershipId}/status:
    patch:
      tags:
        - Tenants
      summary: Update current-tenant membership status
      operationId: updateTenantMembershipStatus
      parameters:
        - name: membershipId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTenantMembershipStatusRequest"
      responses:
        "200":
          description: Updated current-tenant membership.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenantMember"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/tenant-invitations:
    get:
      tags:
        - Tenants
      summary: List user invitations in the current tenant
      operationId: listCurrentTenantInvitations
      description: >
        Returns only invitations for the authenticated active tenant. Requires
        the ManageUsers permission.
      responses:
        "200":
          description: Current-tenant invitations.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TenantInvitation"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      tags:
        - Tenants
      summary: Invite a user to the current tenant
      operationId: createTenantInvitation
      description: >
        Creates a pending invitation with a local notification placeholder. The
        invited role is assigned when the token is accepted. Requires the
        ManageUsers permission and writes an audit event.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTenantInvitationRequest"
      responses:
        "201":
          description: Created current-tenant invitation.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenantInvitation"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          description: A pending invitation already exists for this email in the current tenant.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
  /api/invitations/{token}/accept:
    post:
      tags:
        - Tenants
      summary: Accept an invitation token
      operationId: acceptTenantInvitation
      description: >
        Accepts a pending invitation token, creates an active tenant membership
        with the invited role, and writes an audit event. Expired and revoked
        invitations cannot be accepted.
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
            minLength: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AcceptTenantInvitationRequest"
      responses:
        "200":
          description: Accepted invitation.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenantInvitation"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/tenant-invitations/{invitationId}/expire:
    post:
      tags:
        - Tenants
      summary: Expire a current-tenant invitation
      operationId: expireTenantInvitation
      parameters:
        - name: invitationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Expired invitation.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenantInvitation"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/tenant-invitations/{invitationId}/revoke:
    post:
      tags:
        - Tenants
      summary: Revoke a current-tenant invitation
      operationId: revokeTenantInvitation
      parameters:
        - name: invitationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Revoked invitation.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenantInvitation"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/obligations:
    get:
      tags:
        - Obligations
      summary: List source-backed obligation library records
      operationId: listObligations
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
        - name: q
          in: query
          schema:
            type: string
        - name: source
          in: query
          schema:
            type: string
            examples:
              - FAR 52.204-21
        - name: riskLevel
          in: query
          schema:
            $ref: "#/components/schemas/RiskLevel"
        - name: dataType
          in: query
          schema:
            $ref: "#/components/schemas/DataType"
      responses:
        "200":
          description: Paged obligations.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/PagedResult"
                  - type: object
                    properties:
                      items:
                        type: array
                        items:
                          $ref: "#/components/schemas/Obligation"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/obligations/{obligationId}:
    parameters:
      - $ref: "#/components/parameters/ObligationId"
    get:
      tags:
        - Obligations
      summary: Get an obligation library record
      operationId: getObligationById
      responses:
        "200":
          description: Obligation.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Obligation"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/reports/approved-evidence-packages:
    get:
      tags:
        - Reports
      summary: List approved evidence packages
      operationId: listApprovedEvidencePackages
      description: >
        Returns completed prime evidence package reports in the current tenant
        scope with approved evidence only. Auditor users may call this endpoint
        but cannot use report or evidence modification endpoints.
      responses:
        "200":
          description: Approved evidence packages.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ApprovedEvidencePackage"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Tenant-scoped user access token. MVP implementation may use local development auth before SSO/MFA is enabled.
  parameters:
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: pageSize
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 25
...
    ObligationId:
      name: obligationId
      in: path
      required: true
      schema:
        type: string
      example: far-52-204-21
...
  responses:
    BadRequest:
      description: Request validation failed.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
    Unauthorized:
      description: Missing or invalid authentication.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
    Forbidden:
      description: User is authenticated but not authorized for the requested tenant action.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
    NotFound:
      description: Resource was not found in the current tenant scope.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
  schemas:
    ProblemDetails:
      type: object
      additionalProperties: true
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
        traceId:
          type: string
    PagedResult:
      type: object
      required:
        - items
        - page
        - pageSize
        - totalItems
        - totalPages
      properties:
        items:
          type: array
          items: {}
        page:
          type: integer
          minimum: 1
        pageSize:
          type: integer
          minimum: 1
        totalItems:
          type: integer
          minimum: 0
        totalPages:
          type: integer
          minimum: 0
...
    RoleName:
      type: string
      enum:
        - Owner
        - Admin
        - Compliance Manager
        - Contributor
        - Auditor
        - Advisor
...
    CurrentUserAccess:
      type: object
      required:
        - roles
        - permissions
        - rolePermissionMatrix
      properties:
        roles:
          type: array
          items:
            $ref: "#/components/schemas/RoleName"
        permissions:
          type: array
          items:
            $ref: "#/components/schemas/PermissionName"
        rolePermissionMatrix:
          type: object
          additionalProperties:
            type: array
            items:
              $ref: "#/components/schemas/PermissionName"
    Tenant:
      type: object
      required:
        - id
        - displayName
        - status
        - dataPosture
        - createdAt
      properties:
        id:
          type: string
          format: uuid
        displayName:
          type: string
          maxLength: 240
        status:
          $ref: "#/components/schemas/TenantStatus"
        dataPosture:
          $ref: "#/components/schemas/TenantDataPosture"
        trialEndsAt:
          type:
            - string
            - "null"
          format: date
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type:
            - string
            - "null"
          format: date-time
    TenantMember:
      type: object
      required:
        - membershipId
        - tenantId
        - userId
        - email
        - displayName
        - userStatus
        - membershipStatus
        - roleName
        - mfaEnabled
        - createdAt
      properties:
        membershipId:
          type: string
          format: uuid
        tenantId:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        email:
          type: string
          format: email
          maxLength: 320
        displayName:
          type: string
          maxLength: 200
        userStatus:
          $ref: "#/components/schemas/UserStatus"
        membershipStatus:
          $ref: "#/components/schemas/MembershipStatus"
        roleName:
          $ref: "#/components/schemas/RoleName"
        mfaEnabled:
          type: boolean
        lastSignedInAt:
          type:
            - string
            - "null"
          format: date-time
        lastAccessedAt:
          type:
            - string
            - "null"
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type:
            - string
            - "null"
          format: date-time
    AssignTenantMemberRequest:
      type: object
      required:
        - userId
        - email
        - displayName
        - roleName
      properties:
        userId:
          type: string
          format: uuid
        email:
          type: string
          format: email
          maxLength: 320
        displayName:
          type: string
          minLength: 1
          maxLength: 200
        roleName:
          $ref: "#/components/schemas/RoleName"
        status:
          $ref: "#/components/schemas/MembershipStatus"
    UpdateTenantMembershipStatusRequest:
      type: object
      required:
        - status
      properties:
        status:
          $ref: "#/components/schemas/MembershipStatus"
    TenantInvitation:
      type: object
      required:
        - invitationId
        - tenantId
        - email
        - roleName
        - invitationToken
        - status
        - expiresAt
        - notificationPlaceholder
        - createdAt
      properties:
        invitationId:
          type: string
          format: uuid
        tenantId:
          type: string
          format: uuid
        email:
          type: string
          format: email
          maxLength: 320
        roleName:
          $ref: "#/components/schemas/RoleName"
        invitationToken:
          type: string
          minLength: 1
          maxLength: 128
        status:
          $ref: "#/components/schemas/TenantInvitationStatus"
        expiresAt:
          type: string
          format: date-time
        acceptedAt:
          type:
            - string
            - "null"
          format: date-time
        acceptedByUserId:
          type:
            - string
            - "null"
          format: uuid
        revokedAt:
          type:
            - string
            - "null"
          format: date-time
        revokedByUserId:
          type:
            - string
            - "null"
          format: uuid
        notificationSentAt:
          type:
            - string
            - "null"
          format: date-time
        notificationPlaceholder:
          type: string
          maxLength: 600
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type:
            - string
            - "null"
          format: date-time
    CreateTenantInvitationRequest:
      type: object
      required:
        - email
        - roleName
      properties:
        email:
          type: string
          format: email
          maxLength: 320
        roleName:
          $ref: "#/components/schemas/RoleName"
        expiresInDays:
          type: integer
          minimum: 1
          maximum: 30
          default: 7
    AcceptTenantInvitationRequest:
      type: object
      required:
        - displayName
      properties:
        displayName:
          type: string
          minLength: 1
          maxLength: 200
    CreateTenantRequest:
      type: object
      required:
        - displayName
      properties:
        displayName:
          type: string
          minLength: 1
          maxLength: 240
        status:
          $ref: "#/components/schemas/TenantStatus"
        trialEndsAt:
          type:
            - string
            - "null"
          format: date
    UpdateTenantStatusRequest:
      type: object
      required:
        - status
      properties:
        status:
          $ref: "#/components/schemas/TenantStatus"
    TenantStatus:
      type: string
      enum:
        - Active
        - Trialing
        - Suspended
        - Archived
...
    ComplianceOverview:
      type: object
      required:
        - productPromise
        - mvpDataPosture
        - modules
        - priorityObligations
      properties:
        productPromise:
          type: string
        mvpDataPosture:
          type: string
        modules:
          type: array
          items:
            $ref: "#/components/schemas/ModuleStatus"
        priorityObligations:
          type: array
          items:
            $ref: "#/components/schemas/ObligationSummary"
...
    RiskLevel:
      type: string
      enum:
        - low
        - medium
        - high
        - critical
...
    DataType:
      type: string
      enum:
        - none
        - fci
        - cui
        - cdi
        - classified
        - export-controlled
        - labor-records
        - payroll-records
        - contractor-it
        - supplier-technology
        - unknown
...